real-time-login-tracker/requirements.txt
flask
kafka-python
pymongo
streamlit
